added the authentication to the booking view

+++ littlelemon/restaurant/models.py
@@ -11,4 +11,4 @@ class Menu(models.Model):
     id = models.AutoField(primary_key=True)
     title = models.CharField(max_length=255)
     price = models.DecimalField(max_digits=10, decimal_places=2)
-    inventory = models.IntegerField()
+    inventory = models.SmallIntegerField()
